requirements.txt
flasgger~=0.9.5
Flask~=2.2.2
Flask-Cors~=3.0.10
Flask-JWT-Extended~=4.4.4
psycopg2-binary~=2.9.5
pydantic~=1.10.2
python-dotenv~=0.21.0
SQLAlchemy~=1.4.43
Flask-Pydantic~=0.10.0
argon2-cffi~=21.3.0
Flask-JWT-Extended~=4.4.4
ujson~=5.5.0
humps~=0.2.2